Added visual indication to button that invoices have been created

+++ thinkinglogic-wc-sage-one-integration.php
@@ -105,10 +105,17 @@ function tl_wc_sage_order_actions( $order_id ) {
 		if ($sage->isRefreshTokenExpiringSoon()) {
 			echo '<a class="button" href="' . $sage->authorizationEndpoint() .'">Refresh Authorisation for Sage</a>';
         } else {
-            echo '       <input type="submit"';
-            echo '           class="button button-primary tl_wc_sage_order"';
-            echo '           name="' . ThinkingLogicWCSage::CREATE_INVOICE_BUTTON_ID . '"';
-            echo '           value="Create Sage Invoice(s)"/>';
+			if ($order->meta_exists(self::ORDER_FIELD_CUSTOMER_ID)) {
+				echo '       <input type="submit"';
+				echo '           class="button button-primary tl_wc_sage_order tl_wc_sage_customer_exists "';
+				echo '           name="' . ThinkingLogicWCSage::CREATE_INVOICE_BUTTON_ID . '"';
+				echo '           value="(Re)create Sage Invoice(s)"/>';
+			} else {
+				echo '       <input type="submit"';
+				echo '           class="button button-primary tl_wc_sage_order "';
+				echo '           name="' . ThinkingLogicWCSage::CREATE_INVOICE_BUTTON_ID . '"';
+				echo '           value="Create Sage Invoice(s)"/>';
+			}
         }
 		echo '   </div>';
 		echo '</li>';
